Only download u2net model in docker image
With the addition of the extra models, the docker image is huge (> 17 GB)
and too big to build and push in GitHub actions, especially for `arm64`
(which is slower to build than `amd64`). This change makes it so the `d`
command can take a list of models to download, and the Dockerfile only
downloads the default `u2net` model. This shrinks the docker image to
about 3 GB.

More discussion at: https://github.com/danielgatis/rembg/discussions/735