Stop publishing arm64 image
The arm64 docker image is much slower to build and download models. With
the addition of the extra models, it's now too big to build and push in
Github actions (times out after several hours).

This change removes the arm64 image in favor of just the amd64 image,
which only takes about 9 minutes to build and push.

More discussion at: https://github.com/danielgatis/rembg/discussions/735