remove unneeded log_softmax() call
When generating clothing or background category labels, the log_softmax() call is immediately followed by an argmax() call. In this situation, the log_softmax() is not needed.

Since log_softmax is monotonically increasing, the max value before log_softmax will still be the max value after log_softmax, so we can go to argmax() directly.